requirements.txt
numpy
skfuzzy
pillow
